refactor: add subfolders to storages
