+++ src/indexer/indexer.h
@@ -5,7 +5,7 @@
 #include <string>
 #include "../models/bigToken.h"
 #include "../models/token.h"
-#include "../storages/iIndexStorage.h"
+#include "../storages/indexes/iIndexStorage.h"
 
 
 /**
