+++ src/storages/indexes/iIndexStorage.h
@@ -4,7 +4,7 @@
 #include <vector>
 #include <iostream>
 #include <string>
-#include "../models/bigToken.h"
+#include "../../models/bigToken.h"
 
 /**
  * @brief Interface for a storage index.
